+++ playground/Wallet.tsx
@@ -31,6 +31,10 @@ export const Wallet = () => {
       },
     });
 
+    wallet.isReady.then((readyWallet) => {
+      console.log('Ready wallet (isReady)', readyWallet);
+    });
+
     wallet.subscribe('status-update', (status, prevStatus) => {
       console.log('Status update', { status, prevStatus });
     });
@@ -45,8 +49,13 @@ export const Wallet = () => {
       }
     });
 
-    wallet.subscribe('accounts-update', ([ethAccount, cereAccount]: WalletAccount[]) => {
-      console.log('accounts-update', [ethAccount?.address, cereAccount?.address]);
+    wallet.subscribe('accounts-update', (accounts: WalletAccount[]) => {
+      console.log(
+        'accounts-update',
+        accounts.map((account) => account.address),
+      );
+
+      const [ethAccount, cereAccount] = accounts;
 
       setCereAddress(cereAccount?.address);
       setEthAddress(ethAccount?.address);