+++ src/stores/AssetStore/AssetStore.ts
@@ -1,4 +1,5 @@
 import { makeAutoObservable, autorun } from 'mobx';
+import { getGlobalStorage } from '@cere-wallet/storage';
 
 import { isTransferableAsset, Wallet, Asset, ReadyWallet } from './types';
 import { serializeAssets, deserializeAssets } from './helper';
@@ -18,17 +19,13 @@ export class AssetStore {
   }
 
   async init(wallet: ReadyWallet) {
-    const [{ CereNativeToken }, { NativeToken }, { UsdcToken }, { Erc20Token }, { CereErc20Token }] = await Promise.all(
-      [
-        import(/* webpackChunkName: "CereNativeToken" */ './CereNativeToken'),
-        import(/* webpackChunkName: "NativeToken" */ './NativeToken'),
-        import(/* webpackChunkName: "UsdcToken" */ './UsdcToken'),
-        import(/* webpackChunkName: "Erc20Token" */ './Erc20Token'),
-        import(/* webpackChunkName: "CereErc20Token" */ './CereErc20Token'),
-      ],
-    );
+    const { CereNativeToken } = await import(/* webpackChunkName: "walletAssets" */ './CereNativeToken');
+    const { NativeToken } = await import(/* webpackChunkName: "walletAssets" */ './NativeToken');
+    const { UsdcToken } = await import(/* webpackChunkName: "walletAssets" */ './UsdcToken');
+    const { Erc20Token } = await import(/* webpackChunkName: "walletAssets" */ './Erc20Token');
+    const { CereErc20Token } = await import(/* webpackChunkName: "walletAssets" */ './CereErc20Token');
 
-    const managableTokensFromStorage = localStorage.getItem('tokens');
+    const managableTokensFromStorage = getGlobalStorage().getItem('tokens');
     const parsedAssets: Asset[] = deserializeAssets(managableTokensFromStorage) || [];
 
     this.list = [
@@ -47,7 +44,7 @@ export class AssetStore {
   set managableList(assets: Asset[]) {
     this.managableAssets = assets;
 
-    localStorage.setItem('tokens', serializeAssets(assets));
+    getGlobalStorage().setItem('tokens', serializeAssets(assets));
   }
 
   get loading() {