+++ packages/embed-wallet/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Release Notes:
 
-### vNext
+### v0.8.1
+
+- Fix issue with first connect in mobile browsers
+
+### v0.8.0
 
 - Add CERE ERC20 balance and transfer support
+- Add `isReady` promise property to the wallet instance
+- Fix account changed event delay after wallet connect